requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
qdrant-client
python-dotenv
google-generativeai
pymupdf
pytesseract
Pillow
